requirements.txt
streamlit
numpy
scikit-image
matplotlib
Pillow
torch
torchvision
spotiflow
